requirements.txt
absl-py==2.3.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
docutils==0.21.2
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.58.2
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.72.1
h5py==3.14.0
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
keras==3.10.0
Kivy==2.3.1
Kivy-examples==2.3.1
Kivy-Garden==0.1.5
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numpy==2.1.3
omegaconf==2.3.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
Pygments==2.19.1
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
thop==0.1.1.post2209072238
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.151
ultralytics-thop==2.0.14
ultralytics-yolov8==0.0.1
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
